fix: populate image src
