+++ apps/renderer/src/components/ui/markdown/renderers/BlockImage.tsx
@@ -4,6 +4,7 @@ import { useWrappedElementSize } from "~/providers/wrapped-element-provider"
 import { useEntry } from "~/store/entry"
 
 import { Media } from "../../media"
+import { usePopulatedFullUrl } from "../utils"
 
 export const MarkdownBlockImage = (
   props: React.ImgHTMLAttributes<HTMLImageElement> & {
@@ -14,6 +15,9 @@ export const MarkdownBlockImage = (
   },
 ) => {
   const size = useWrappedElementSize()
+  const { feedId } = useEntryContentContext()
+
+  const src = usePopulatedFullUrl(feedId, props.src)
 
   const { entryId } = useEntryContentContext()
   const media = useEntry(entryId, (entry) => entry?.entries.media?.find((m) => m.url === props.src))
@@ -22,6 +26,7 @@ export const MarkdownBlockImage = (
     <Media
       type="photo"
       {...props}
+      src={src}
       height={media?.height || props.height}
       width={media?.width || props.width}
       blurhash={media?.blurhash}
